requirements.txt
fastapi
uvicorn
openai
python-dotenv
slowapi
sqlalchemy
httpx
python-multipart
